refactor: add xarray submodule and consistent naming between xarray and array
