+++ src/earthkit/meteo/score/__init__.py
@@ -7,13 +7,5 @@
 # nor does it submit to any jurisdiction.
 #
 
-"""
-Verification functions.
-
-The API is split into two levels. The low level functions are in the ``array`` submodule and they
-can be used to operate on numpy arrays. The high level functions are still to be developed and
-planned to work with objects like *earthkit.data FieldLists* or *xarray DataSets*.
-"""
-
-from .deterministic import *
-from .ensemble import *
+from .xarray.deterministic import *
+from .xarray.ensemble import *
